♻️ rfac(custom): project 타입 분리
