+++ src/apis/projects/post.ts
@@ -1,13 +1,8 @@
+import { Project } from '@/types/project';
+
 import { typedPost } from '..';
 
-export const addProject = async (body: AddProjectRequest) => {
+export const addProject = async (body: Project) => {
   const response = await typedPost('/project', body);
   return response;
 };
-
-interface AddProjectRequest {
-  title: string;
-  startDate: string;
-  endDate: string;
-  goal: string;
-}
