+++ src/apis/projects/put.ts
@@ -1,14 +1,8 @@
+import { Project } from '@/types/project';
+
 import { typedPut } from '..';
 
-export const editProject = async (id: number, body: EditProjectResponse) => {
+export const editProject = async (id: number, body: Project) => {
   const response = await typedPut(`/projects/${id}`, body);
   return response;
 };
-
-interface EditProjectResponse {
-  title: string;
-  startDate: string;
-  endDate: string;
-  goal: string;
-  content: string;
-}
