+++ src/types/project.ts
@@ -1,0 +1,7 @@
+export interface Project {
+  title: string;
+  goal: string;
+  content: string;
+  startDate: string;
+  endDate: string;
+}
